requirements.txt
fastmcp
python-dotenv
git+https://github.com/natalieee-n/bondage-club-bot-core.git
